+++ TugasBesarPBO/nbproject/genfiles.properties
@@ -1,8 +1,8 @@
-build.xml.data.CRC32=cdae8e42
+build.xml.data.CRC32=9451612e
 build.xml.script.CRC32=576a7452
-build.xml.stylesheet.CRC32=8064a381@1.75.2.48
+build.xml.stylesheet.CRC32=f85dc8f2@1.95.0.48
 # This file is used by a NetBeans-based IDE to track changes in generated files such as build-impl.xml.
 # Do not edit this file. You may delete it but then the IDE will never regenerate such files for you.
-nbproject/build-impl.xml.data.CRC32=cdae8e42
-nbproject/build-impl.xml.script.CRC32=66e84407
-nbproject/build-impl.xml.stylesheet.CRC32=876e7a8f@1.75.2.48
+nbproject/build-impl.xml.data.CRC32=9451612e
+nbproject/build-impl.xml.script.CRC32=6e221ec0
+nbproject/build-impl.xml.stylesheet.CRC32=f89f7d21@1.95.0.48
