+++ TugasBesarPBO/nbproject/private/private.properties
@@ -5,4 +5,4 @@ do.jlink=false
 javac.debug=true
 javadoc.preview=true
 jlink.strip=false
-user.properties.file=C:\\Users\\dense\\AppData\\Roaming\\NetBeans\\8.0.2\\build.properties
+user.properties.file=C:\\Users\\Win10\\AppData\\Roaming\\NetBeans\\12.0\\build.properties
