+++ TugasBesarPBO/nbproject/project.properties
@@ -73,7 +73,7 @@ main.class=View.Main
 manifest.file=manifest.mf
 meta.inf.dir=${src.dir}/META-INF
 mkdist.disabled=false
-platform.active=default_platform
+platform.active=JDK_8.231
 run.classpath=\
     ${javac.classpath}:\
     ${build.classes.dir}
