+++ TugasBesarPBO/src/Model/Users.java
@@ -98,7 +98,8 @@ public class Users {
     public boolean cekLogin(String username, String password){
         DatabaseControl controller = new DatabaseControl();
         Users user = controller.getUser(username);
-        if((user.getUsername().equals(username)) && (Convert.MD5ToString(user.getPassword()).equals(password))){
+        String passwordConvert = Convert.stringToMD5(password);
+        if((user.getUsername().equals(username)) && (passwordConvert.equals(user.getPassword()))){
             UserManager.getInstance().setUser(user);
             return true;
         }else{
